refactor : Controller 패키지와 Model 패키지로 구분
